Merged PR 19: Updating Method names
There are quite a few changes in this one, but no real change to functionality, really only changing method names to match java naming conventions. Also changed some method/attribute/class names to make them make a little more sense with what they are doing.

Not going to link a work item because it kind of is a bunch but none of them at the same time.