Merged PR 23: Level classes and Level creation process
This PR creates adds a few classes for the process of generating a gameBoard from data for a Level, and the general Level control structure in general.

There is a lot here, I plan to try to add some documentation to the wiki later today.

Also probably still some tweaks and changes to be made, but this should at least be working for now

Related work items: #22