Merged PR 37: Created the function to make the collision alert pop up.
There is currently functionality to make the alert pop up, but not a place to make the alert pop up. As soon as there is some sort of indication of an obstacle being hit, it will be as simple as a few lines of code to actually get this implemented.

Related work items: #70