Merged PR 45: level2-and-reset
I added the level 2 data to the LevelData .txt file. I couldn't quite figure out what to do with the special commands even though I had plans for customizing those too. I'm sure it will be easy to tweak once I ask about it.

The level reset function is in the LevelController class, and it resets the Roomba's starting position and command list. I could also code it to reset collectibles if that's needed right now.

Message me if there's anything wrong with the code I wrote and I will try to message back as soon as I can.