Merged PR 88: Populate move Lists when saved attempt is clicked
Change the format of the string for saving the move lists, and then repopulate moveLists when saved attempt is clicked.

There is a bug where clicking on a move item after lists have been repopulated doesn't select it right away, have to click a moveItem to "activate it", then click a different one, then click original again and it will select.

Related work items: #202, #205