Merged PR 91: New SFX, ability to change sfx volume in settings.
I've removed the old sfx and replaced them with 3 new ones, which should cover the essential sfx features. I changed the settings menus so there's only one seekbar for all sound effects. The seekbar is fully functional.

Also, the icons for collectibles and empty spaces are wrong when playing. Will that cause problems when merging?